feature: добавить валидацию, обработать ошибки

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -4,8 +4,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.practicum.shareit.user.dao.UserDao;
 import ru.practicum.shareit.user.dto.UserDto;
+import ru.practicum.shareit.user.exception.EmailAlreadyExistException;
 import ru.practicum.shareit.user.mapper.UserMapper;
 import ru.practicum.shareit.user.model.User;
+import ru.practicum.shareit.util.exception.NotFoundException;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -15,30 +20,68 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDto addNewUser(UserDto userDto) {
-        User user = new User();
+        checkEmailDuplicate(userDto.getEmail());
 
-        userDao.save(user);
+        User user = userMapper.toUser(userDto);
+        User userCreated = userDao.save(user);
 
-        return userMapper.toUserDto(user);
+        return userMapper.toUserDto(userCreated);
     }
 
     @Override
     public UserDto getUserById(long id) {
         User user = userDao.findById(id);
+
+        if (user == null) {
+            throw new NotFoundException("пользователь с id " + id + " не найден");
+        }
+
         return userMapper.toUserDto(user);
     }
 
     @Override
     public UserDto updateUser(long id, UserDto userDto) {
-        User user = new User();
+        User user = userDao.findById(id);
 
-        userDao.update(id, user);
+        if (userDto.getName() != null) {
+            user.setName(userDto.getName());
+        }
+        if (userDto.getEmail() != null && !user.getEmail().equals(userDto.getEmail())) {
+            checkEmailDuplicate(userDto.getEmail());
+        }
+        if (userDto.getEmail() != null) {
+            user.setEmail(userDto.getEmail());
+        }
 
-        return userMapper.toUserDto(user);
+        return userMapper.toUserDto(userDao.update(user));
     }
 
     @Override
     public void deleteUserById(long id) {
+        User user = userDao.findById(id);
+
+        if (user == null) {
+            throw new NotFoundException("пользователь с id " + id + " не найден");
+        }
+
         userDao.deleteById(id);
+    }
+
+    @Override
+    public List<UserDto> getAllUsers() {
+        List<User> users = userDao.findAll();
+        List<UserDto> usersDtos = new ArrayList<>();
+
+        for (User user : users) {
+            usersDtos.add(userMapper.toUserDto(user));
+        }
+
+        return usersDtos;
+    }
+
+    private void checkEmailDuplicate(String email) {
+        if (userDao.emailExist(email)) {
+            throw new EmailAlreadyExistException("Такой email уже существует");
+        }
     }
 }
